feat(bdd): extract supabase mock factory for reuse in cucumber world
Creates supabase-factory.ts with createSupabaseMock() that accepts an
optional spyWrap parameter, allowing Vitest (vi.fn) and BDD (plain fn)
contexts to share the same mock builder logic. Refactors supabase.ts to
delegate to the factory.
