+++ src/test/mocks/supabase.ts
@@ -1,45 +1,6 @@
-export const mockQueryResult: { data: unknown; error: unknown } = { data: null, error: null };
+import { vi } from 'vitest';
+import { createSupabaseMock } from '@/test/mocks/supabase-factory';
 
-const methods = [
-  'from',
-  'select',
-  'insert',
-  'update',
-  'delete',
-  'upsert',
-  'eq',
-  'neq',
-  'is',
-  'not',
-  'or',
-  'order',
-  'single',
-  'maybeSingle',
-  'limit',
-  'range',
-  'contains',
-  'match',
-  'gte',
-  'lte',
-  'gt',
-  'lt',
-  'like',
-  'ilike',
-  'in',
-  'textSearch',
-] as const;
+const mock = createSupabaseMock(((fn) => vi.fn(fn as never)) as Parameters<typeof createSupabaseMock>[0]);
 
-type MockBuilder = Record<string, ReturnType<typeof vi.fn>> & {
-  then: (resolve: (v: unknown) => void, reject?: (e: unknown) => void) => Promise<unknown>;
-};
-
-const builder = {} as MockBuilder;
-const self = () => builder;
-
-for (const m of methods) {
-  builder[m] = vi.fn(self);
-}
-
-builder.then = (resolve, reject) => Promise.resolve({ ...mockQueryResult }).then(resolve, reject);
-
-export const supabase = builder;
+export const { supabase, mockQueryResult } = mock;
